+++ webapp/components/form/InputChips/index.js
@@ -1,0 +1,1 @@
+export { default } from './InputChips'
