+++ webapp/components/form/InputChips/store/index.js
@@ -1,0 +1,2 @@
+export { useInputChips } from './useInputChips'
+export { State } from './state'
